+++ src/translations/en.json
@@ -6885,7 +6885,6 @@
           "home": {
             "summary_list": {
               "climate": "Climate",
-              "lights": "Lights",
               "security": "Security",
               "media_players": "Media players"
             },
